Standardizing code base strings on double quotes instead of a mix.
This change was made in response to issue #18 in the repo. It's an
aesthetic change and should have no effect on the actual underlying code.
In addition, small fixes were made to get linting to 10/10, including
adding a docstring to the quince.utility module and quince.components.deck
line 59 (comment was above character limit for one line).